chore: move UIfmt to foundry-common
