+++ cast/Cargo.toml
@@ -12,6 +12,7 @@ keywords = ["ethereum", "web3"]
 foundry-utils = { path = "../utils" }
 foundry-evm = { path = "./../evm" }
 foundry-config = { path = "./../config" }
+foundry-common = {path = "./../common"}
 
 futures = "0.3.17"
 ethers-etherscan = { git = "https://github.com/gakonst/ethers-rs", default-features = false }
