+++ common/Cargo.toml
@@ -15,7 +15,6 @@ foundry-config = { path = "../config" }
 
 # eth
 ethers-core = { git = "https://github.com/gakonst/ethers-rs", default-features = false }
-
 # cli
 clap = { version = "3.0.10", features = [
     "derive",
@@ -26,4 +25,5 @@ clap = { version = "3.0.10", features = [
 
 #  misc
 serde = "1.0.133"
+serde_json = "1.0.67"
 thiserror = "1.0.31"
